requirements-load.txt
google-cloud-bigquery==2.2.0
google-cloud-storage==1.32.0
pyyaml==5.3.1
